OMAP: PM: replace omap_pm_get_dev_context_loss_count with omap_pm_was_context_lost
Drivers just need to know if thier context was lost
or not. There is not much point in maintaining a count

This removes omap_pm_get_dev_context_loss_count and
replaces it with omap_pm_was_context_lost

Adapt drivers to use the omap_pm_was_context_lost API,
which returns a bool value instead of relying on the
incrementing context lost counter. This makes
drivers a little more snappier as they do not need to
restore contexts when not needed.

WARNING: this changes the api set available to drivers
to use from OMAP PM layer.

Change-Id: I7fd9183d3e12982bb40651df9378637073121399
Signed-off-by: Axel Haslam <axelhaslam@ti.com>