OMAP: PM: use context lost register to restore context.
Add the logic in omap_pm_was_context_lost so that it
reads the context lost register to determine if
context was lost in the module and return true or
false accordingly.

Handle the special case of ABE, where, logic registers
are always restored by the driver, and constant reload
causes problems, hence only report context lost when
mem is lost.

Change-Id: I39f5e74a0c0e33fbe1d498314a891eb4ec0073e6
Signed-off-by: Axel Haslam <axelhaslam@ti.com>