OMAP4: PM: disable IVA AUTO_RET
OMAP4 silicon has a under-investigation bug in PRCM identified on a
product line concerning the transitions on IVA voltage domain during
retention stage. The behavior of the bug seems to that the PRCM
state machine starts transition into retention path when
AUTO_RET is set, however, the resultant voltages seen on vdd_iva seems
to indicate that VC has been commanded by the vfsm to go to low-power,
ret/off voltages. The resultant behavior for IVA is unpredictable.
In the case of OFF transitions, the transition from this un-defined
state to OFF could even result in hang of the silicon.

It is hence been advised to disable AUTO_RET feature of PRCM VC module
for IVA voltage domain on OMAP4(4430 and 4460) as a stability workaround
untill a more optimal workaround is identified or fix provided.

This has the following impact:
for OSWR suspend, idle: IVA voltage does not hit retention voltage.
for OFF mode: IVA transitions directly from ON to OFF and does not
enter into RET path.
Maximum power impact as a result of this change is expected to be in the
range of 1-2mA.

As part of this change, introduce pm_errata handling for PM44xx
which will allow us to scale errata handling based on silicon revision
on the fly.

Change-Id: I206be6e12a8cd234030337220470f25e6721ecfb
Signed-off-by: Nishanth Menon <nm@ti.com>