drivers: remoteproc: print traces in a circular way
Previosly the traces were printed always from the beginning of the buffer trace
till the end causing confusions when trying to identify the last one.

Now remoteproc is writing a write_idx and read_idx at the end of the trace
buffer to keep track of the traces in the buffer.

We read the indices and and print from the read idx to the end and from the
beginning of the trace buffer to the write idx.  With this we will see the
traces in order from the oldest to the newest ones.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>