i2c: omap44xx: disable ARDY interrupt for each transfer
For indicating ready-for-next-transfer, there is ARDY interrupt.
There is no need to have this interrupt enabled for indication of current
transfer as complete.

So resort to a combination of Interrupt and Polling mechanism.
Interrupt mechanism for: RRDY/RDR and XRDY/XDR
Polling mechanism for: ARDY

Before a message is to be sent, check for ARDY status ready with a timeout protection.

For one message the flow is:
 Fifo operation; one interrupt; Done
 Thus avoiding ARDY wait/interrupt altogether.

For two/more message transfer the flow is:
 Fifo operation;one interrupt; wait_ardy(); fifo operation;Done
 Thus avoiding 2 ARDY interrupts, instead only one ardy wait

Change-Id: Icf26731db373073bc300da2804979b5d509ac448
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>