usb: musb: omap2430: process all otg events, not just the latest one
Currently, the omap2430 musb driver just caches the latest otg
event and schedules work in its otg notifier. Because of this,
it may miss intermediate events and only handle the latest one.
This can cause the driver state to become inconsistent since
it doesn't store whether or not it has enabled the phy, or
did a runtime_pm get, but relies purely on the fact that events
occur in the proper order.

This patch converts the single work item in musb struct into
a singlethreaded workqueue, to which work items for each otg event
are allocated and posted. This forces the driver to process every
event in order.

Ideally, the driver should keep the state of what is enabled so that
this is not necessary.

Change-Id: I123e40d62424a0e339549e6ff051b423f24c4d08
Signed-off-by: Dima Zavin <dima@android.com>