ASoC: ABE: Move the enabling of the IO task for the Frontends.
IO tasks for FE and some Back end are enabled too early. It needs
to be done only on the enable_data_transfer function. This will
avoid to have ABE ATC started before the stream. This patch is
needed in order to remove the pop on start of any FE.

Change-Id: Id898552742703ba67874754e1a2684fdcb0eb7b7
Signed-off-by: Sebastien Guiriec <s-guiriec@ti.com>
Signed-off-by: Pankaj Jindal <pankajjindal@ti.com>
Signed-off-by: Misael Lopez-Cruz <misael.lopez@ti.com>