ASoC: ABE: Add digital mute for FEs
Digital mute for ABE FEs is not exclusively for the FE itself,
but with respect to active BEs. For example, digital mute for
MM_DL can be the MM_DL->DL1 Mixer gain if PDM_DL1 is a client
of MM_DL FE. Similarly for MM_DL->DL2 Mixer gain is PDM_DL2 is
another client of MM_DL FE.

Due to the above relationship between ABE FEs and BEs in terms
of digital mute, FE digital mute has to handle static and
dynamic route changes.

Change-Id: Ib6d05cda2397d3511cd083e4af5d64dcdba34ac6
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>