OMAP4:HSI: Fix CA_WAKE IRQ loss in L3INIT OSWRET
In Open Switch Retention, the asynchronous wake up of the HSI IP
is not possible from the CA_WAKE signal.

Open Switch Retention is possible in Idle mode, so we forbid the idle
mode in this patch when we need it.

In details:
When CAWAKE goes low with data still in RX FIFO, HSI driver tries
to disable the clocks, but aborts because FIFO is not empty. So IO
Daisy chain is not enabled.

Since CAWAKE is low, HSI driver sets SYSCONFIG to SMART_IDLE_WAKEUP
and SMART_STBY_WAKEUP, so L3INIT_PD might be powered down. Since
IO daisy chain has not been enabeld, we miss the following CAWAKE
high event.

Fix consists in always forcing SYSCONFIG to NO_IDLE while there
is any HSI activity, and let omap_hwmod re-enable the
SMART_WAKEUP features as soon as HSI doesn't need the clocks anymore.

Change-Id: I86994aafa4701e549faa52e9d8648f6bd4a464ef
Signed-off-by: Cedric Poignet <a0919860@ti.com>
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>