OMAP4: DVFS: Adding volt request for dep domain should be based on user value
The volt request for dependent domain should be added based on the user
requested value but _not_ from the plist of the voltdm.

The issue is seen during A/V playback, rpres is not releasing constraint on CORE once
the A/V playback is exited.
There is a corner case issue, when two req_dev which share the same voltdm does device_scale.
And that voltdm has dependency on other voltdm. In this scenario,
one of the req_dev will end up not releasing the request from the _dependent_ volt domain.

For example: See the below log captured after A/V playback

==> iva/info <==
vdd_iva
|- voltage requests
|  |
|  |-950000: rpres:iva.0
|  |-950000: aess:aess
|  X
|
|- frequency requests
|  |
|  |- rpres:iva.0
|  |  |-124160000: rpres:iva.0 --> rpres freq request for iva at OPP50
    |  |  |-124160000: aess:aess ---> AESS requested dependency at OPP50
    |  |  X
    |  |- platform:dsp.0
    |  |  `-none
    |  |- aess:aess
    |  |  |-98304000: aess:aess --> AESS at OPP50
    |  |  X
    |  X

==> dep core/info <==
    vdd_core
    |- voltage requests
    |  |
    |  |-962000: :omap_pm_set_min_bus_tput
    |  |-962000: platform:mpu.0
    |  |-962000: pvrsrvkm:pvrsrvkm.0
    |  |-962000: aess:aess
    |  |-1127000: rpres:iva.0 ---> Still at OPP100 volt. This is the result of the bug!!!
    |  X
    |
    |- frequency requests
    |  |
    |  |- platform:l3_main_1.0
    |  |  |-100000000: :omap_pm_set_min_bus_tput
    |  |  |-100000000: platform:mpu.0
    |  |  |-100000000: pvrsrvkm:pvrsrvkm.0
    |  |  |-100000000: aess:aess
    |  |  |-200000000: rpres:iva.0
    <snip>

In the above scenario, i.e. in a/v usecase, rpres and aess both request device_scale
of OPP100. As both are in IVA voltdm, it will add user request for IVA. And as IVA
has dependency on CORE, there is indirect user request (rpres and aess) added on CORE
as well. Now, say rpres releases the constraint by requesting OPP50, and assume aess
didn't release the request yet. At this point, device_scale will update rpres user
request for IVA voltdomain to OPP50. However, for dependent domain it will still end
up requesting OPP100. Because the dependent domain volt request is taken from the plist.
And in the plist for IVA voltdm the higest volt is still OPP100, as aess is still active.
So, this will end up rpress to have OPP100 request on CORE voltdm after A/V usecase.

Fix:
Move the scaning of dependent volt at the early stage in device_scale and update the
request for dependent domain with user requested rate and not the plist value.

Change-Id: Ie3b708acacc9926ec6bb4b9e03a004e5a2432385
Signed-off-by: Girish S G <girishsg@ti.com>