ARM: omap4: tuna: change i2c4 SDA line load
i2c4 has 8 devices attached.
HW recommendation is to have minimum R*C value so that SDA drift is minimised.

C is fixed for the board with the 8 devices attached.
Expermentally its proved that 2.2K external pull is enough for i2c.4 load.
So Internal Pull of i2c.4 should be completely disabled.

Change-Id: I32d187aa3e554a6e2e1a4a87f84129e880313605
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Amancha Vinay <vinaykumar@ti.com>