OMAP4: PM: syscontrol: workaround LPDDR CK IO pull disable
CK differential IO PADn, PADp values can be same because of a hardware bug in
EMIF CMD phy. To work around this bug, it is necessary to disable the
pulls. Though this bug has no functional impact, in OSWR, 500uA saving is
expected with the workaround for these IOs.

CDDS no: OMAP4460-1.0BUG00291 (OMAP official errata ID yet to be available).

Change-Id: I5a5df9ee7a2037279778cf9fab0aca1443222aa8
Signed-off-by: Ambresh K <ambresh@ti.com>
Signed-off-by: Vinay Amancha <vinaykumar@ti.com>