OMAP3+: HSI: mark runtime device to be used in irq_safe and re-enable run-time-pm use
Commit 3bddebb31b924deba279ca757dd5e3ebef4c65ba
(omap_hsi: HACK: Don't use runtime pm) disabled runtime pm
usage in HSI driver. On further investigation, it was found
that runtime API's are getting called from atomic context.

Mark them as IRQ safe as documented in runtime documentation,
drivers can call put_sync in contexts where sleep is allowed,
in contexts where sleep is not possible, the drivers need to
mark these as to be made interrupt safe and also use put
sync_suspend instead of put_sync as the idle callbacks will
be called with interrupt enabled - which is not a good thing
to happen in ISR/interrupt disabled context. Two possibilities
come to mind:
a) If the interrupts get enabled during idle, you can miss an
   timer interrupt
b) The biggest risk is a deadlock when somebody has a spinlock
   held with IRQs turned off, then accidentally turns IRQs on,
   and an interrupt fires and tries to take the same spinlock

With a verification patch such as https://patchwork.kernel.org/patch/1001572/,
we see the following. This patch fixes it.
[   15.016723] [<c0052954>] (dump_backtrace+0x0/0x10c) from [<c0561c60>] (dump_stack+0x18/0x1c)
[   15.025848]  r7:c6d52000 r6:ffff717d r5:c0718080 r4:c6d52000
[   15.032348] [<c0561c48>] (dump_stack+0x0/0x1c) from [<c0080698>] (__might_sleep+0x108/0x128)
[   15.041351] [<c0080590>] (__might_sleep+0x0/0x128) from [<c0295148>] (__pm_runtime_resume+0x90/0x98)
[   15.051147]  r6:c781f000 r5:00000004 r4:c787ba08
[   15.056457] [<c02950b8>] (__pm_runtime_resume+0x0/0x98) from [<c03c0fec>] (hsi_clocks_enable_channel+0x30/0x34)
[   15.067199]  r7:c6d52000 r6:c781f000 r5:c781f968 r4:c787ba08
[   15.073669] [<c03c0fbc>] (hsi_clocks_enable_channel+0x0/0x34) from [<c03c31e4>] (do_hsi_tasklet+0x38/0xfc)
[   15.084014]  r5:c781f968 r4:c781f000
[   15.088073] [<c03c31ac>] (do_hsi_tasklet+0x0/0xfc) from [<c0096af0>] (tasklet_hi_action+0x128/0x188)
[   15.097869]  r8:c0728598 r7:c6d52000 r6:00000000 r5:c781f488 r4:c781f484
[   15.105529] [<c00969c8>] (tasklet_hi_action+0x0/0x188) from [<c0097054>] (__do_softirq+0xb0/0x178)
[   15.115081] [<c0096fa4>] (__do_softirq+0x0/0x178) from [<c009738c>] (irq_exit+0x98/0xa0)
[   15.123870] [<c00972f4>] (irq_exit+0x0/0xa0) from [<c0048058>] (asm_do_IRQ+0x58/0xb4)
[   15.132324]  r5:c0046e98 r4:c0728598
[   15.136444] [<c0048000>] (asm_do_IRQ+0x0/0xb4) from [<c004e448>] (__irq_svc+0x48/0xe0)
[   15.145050] Exception stack(0xc6d53c88 to 0xc6d53cd0)

Change-Id: I0816fd9c1b94002827cf72153ebac25fe2b7ec0d
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>