OMAP:DSS:DSSCOMP: Workaround for displaying garbage on screen.
It seems that DISPC programming may be unsuccessful even when
we see the GO bit cleared.  Delay unpinning TILER blocks for the
just displayed frame because it still may be displaying (if
the programming of the next frame was unsuccessful.)

Obviously, there is a danger of someone writing to these frames
as they are released for writing even if we don't unpin them, but
most of the time this is not happening, so this patch will provide
better visual performance.

Change-Id: I4f94ffad71d90e7dea8a603b05d1c95985debeea
Signed-off-by: Lajos Molnar <molnar@ti.com>