OMAP4: HSI: Fix missign break; setting HSI in 3-wires mode
There is a missing break in the SET_ACREADY_NORMAL IOCTL, so every time we use
the HSI_IOCTL_SET_ACREADY_NORMAL, the HSI is set in 3 wires modes, which forces the
ACREADY to default to 1.
This is why we observed a frame being received while ACREADY doesn’t go back to 0.

This was causing ACK timeout issue.

Change-Id: I20accfa258810fb836667d333555456753df46d1
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>
Reported-by:wonhee <wonhee48.seo@samsung.com>