i2c: omap: recover from Bus Busy condition
In case a peripheral is holding the DATA bus low, provide a 400Khz
constant clock output using the TEST register.

Also soft reset the I2C controller so that there is no stale state
left in the HW state machine.

A WARN_ON() will be generated when a BB timeout happens.

Change-Id: I23b824f6ae3435ea2e4abf0cbc047584b1e9b449
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>