rpres: dev: put ivahd under reset after it is released
Sometimes the remote processor crashes leaving the ivahd with
pending transitions then the rpres module put it into idle but
the ivahd cannot go to idle and ended up in transition state.

Since the ivahd does not have softreset call omap_device_shutdown
instead of omap_device_idle, that way it is put under hardreset
and it is functional when other request is done since the interrupts
are cleared.

Change-Id: I2211bb7d37931b6c795fda0fbab664061b2616d0
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>