OMAP4+: EMIF: make interrupts wakeup capable
EMIF has multiple notifiers on it's interrupt line. The one that is
interesting is the temperature alert. On POP devices such as OMAP4,
the LPDDR (to which EMIF is interfaced) is in juxtaposition to OMAP
where thermal characteristics closely (if not precisely) follow that
of OMAP. On high performance silica such as OMAP4460, with hotspot
closer to MPU, when MPU goes to higher frequencies, LPDDR follows
the temperature events by requesting MPU over EMIF interrupt for
the standard events such as nominal, derate_refresh, derate_refresh
and_timing or even shutdown requests.

When MPU enters suspend from very high frequency operation into
a low power state such as OSWR/OFF mode, the timing of the interrupt
from EMIF(which is driven by core domain) for events such as nominal
temperature alert(from high derate) could be very very close in time
to MPU completing the wfi instruction and hitting requested low power
state. If the event occurs before MPU hits wfi instruction, MPU
never enters the low power state. However, when the event occurs
after MPU hits the low power state, since the EMIF interrupt is not
configured to be wakeup capable, WUGEN never wakesup MPU and we have
a pending interrupt.

In this scenario, since EMIF is driven from core, due to the pending
event, core never achieves the low power state and we have a situation
where MPU and other domains have hit required power states, core
power domain alone does not hit low power state due to the pending
event. The system then needs to wait for an interrupt which is
actually marked as a wakeup interrupt to wake MPU and process both
the events (the wakeup event and the pending EMIF interrupt).

Since EMIF events are important enough to be processed by MPU ASAP
and the fact that this can prevent system level power state achievement,
make the EMIF interrupts as wakeup capable.

Change-Id: I3899fac187ec37519b97c37942506dd96ffac0a7
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Arojit Roychowdhury <aroyc@ti.com>
Signed-off-by: Vinay Chaurasia <vinaych@ti.com>