OMAP4: PM: dont suspend if any devices were scaling
If we interrupt an device scale operation, it is possible that
we might have caught the operation off-guard such as in the middle
of voltage scale or even frequency change operation. we might be
unable to guarentee that all cross domain dependencies are taken
care as well. Instead, we will abort suspend and priortize device
scale operation and allow the system to re-attempt suspend when
the conditions are right for the operation to proceed.

Change-Id: I474a1d3bba0e8fcab7e2524cbd015d87bf84ce9d
Signed-off-by: Nishanth Menon <nm@ti.com>