OMAP4: PM: Fold saving power state with programming suspend state
There are two loops as part of going to suspend state. The
first walk through saves the power states and the second loop
programs the LP state as part of going to suspend. Fold
them to a single loop where the state is saved followed by
programming the required power state.

Change-Id: Ibc1a0cff2c643f7b23d07af92b8ab5fc291e3dd9
Signed-off-by: Nishanth Menon <nm@ti.com>