OMAP4: PM: Dont write to readonly/reserved powerdomain register on resume
Some of the powerst control/logic_st_ctrl registers such as
CAM, EMU, ABE, etc. are reserved registers, we shouldn't be
writing to them in the first place. The ALWAYS_ON power domains
(wakeup, those of core, DSP, MPU), STD_EFUSE are not under
software control, hence for the purposes of this patch, considered as
not modifiable by software. We need to honor this as part
of resume restore of power domain as well.

Change-Id: Ib3e4a3fc245d3ba18dfad18d17920fcf79b85cf0
Signed-off-by: Nishanth Menon <nm@ti.com>