usb: musb: Handle short packet for mode1 RX path
This patch adds the missing logic to handle a
short packet when recieved in DMA mode1 RX.
If the DMA is programmed in mode1 RX and the
data recieved is less than the programed DMA count,
we need to abort the dma xfer on recieving a short
packet (which marks the end of the transfer), switch
to mode0 and service the short packet.

This logic is needed only for MTP/PTP case and breaks
Mass-storage since Mass-storage doesnt require this
additional handling of short packet, hence the
short_not_ok flag (set only for mass-storage)
is used as check for not executing this logic for
mass-stoarge gadgets.

Tested: MTP(Windows Host)/PTP/ADB(UBUNTU Host) using
differnt size files between 512B-500MB.

Change-Id: Ieb84bcccfcf4e7d951378bef4d9f1dc6dd5a45fd
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>