OMAP4: HSI: PM: Errata: DSP woken-up by HSI CAWAKE signal
Description:
There is a silicon bug in HSI IP.
The logic for generating the SWakeup for DSP is same as the logic for generating
SWakeup for MPU, even when DSP interrupt generation is disabled.
As a result when the MPU Wakeup is generated, a false DSP wakeup is also
generated.

Consequences of this bug in below scenario:
- Modem interface (HSI is enabled)
	- The system is allowed to suspend
	- When CAWAKE signal has activity, it will wake-up MPU; it also wakes up
	DSP incorrectly
	- The DSP is not able to handle this wake up correctly.
	- Next suspend, system can’t enter in off-mode because DSP remains
	active

This errata needs to be handled in two scenarios:
a) If the device has woken-up from RET/OFF through CAWAKE signal and is taken
care as part of wakeup from suspend.
b) During MPU active and HSI transitioning from its internal low power state,
will generate SWakeup signal which eventually wakeup's DSP. This prevents
DSP (Tesla) from entering low power state and thus handling this corner case
in HSI interrupt handler as recommended by silicon team.

Signed-off-by: Vinay Chaurasia <vinayc@ti.com>
Signed-off-by: Ambresh K <ambresh@ti.com>