usb: ehci: omap4xx: fix resume failure
It has been seen that there is noise during the 3ms no
activity period on the ULPI lines. This causes sometimes the Port Resume
not to work.

So the SW workaround is to reset the EHCI and PHY and restart the RESUME signalling.
By resetting the EHCI, the state m/c is in good state and RESUME will again work.

If for some reason, RESUME did not work first time, then resort to a full re-enumeration.

Change-Id: I02eb2552a890f67027f1eb99b756fdae32a9f02d
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>