OMAP: DSS/DSSCOMP: Wait until manager blank applies
Before turning off a display, wait until we had a chance to turn off
all overlays on the manager.  Otherwise, DISPC may get into an undefined
state if we are trying to reuse those overlays on another manager.

Also, turning the display off while the GO bit is set may be an issue
even if the overlays are not reused on another manager.

Change-Id: I43daa4603ec10617b26e8ad2be88f2a25db6c1df
Signed-off-by: Lajos Molnar <molnar@ti.com>