ARM: omap4: tuna: use androidboot.mode=charger parm to enable fuel alert
The 17043 has an interrupt that shuts the system down if the soc is
less than 2%.  This interrupt should only be enabled in normal
boots. It should not be enabled if we are booting to charger mode.

The bootloader passes the parameter "androidboot.mode=charger" to tell
userland to boot to charger mode.  This parameter is currently unused
in the kernel.

Add code to the power board file to set a flag to 0 in normal mode and
1 in charger mode.  This flag will be used in the 17043 driver and
passed to it through its platform data.

Change-Id: If040b93c4fef6024890ad6b95b5875bffc75f753
Signed-off-by: Marty Fouts <mfouts@sta.samsung.com>