remoteproc: introduce rproc method dump_registers
This patch introduces method dump_registers() in rproc_ops and replaces the
default function dump_remoteproc_regs() with one implemented by each remoteproc
(if defined.)  Function dump_remoteproc_regs() called __show_regs(), which
attemps to print the registers given to it as if they belong to local SMPs
and further attempts to dump memory around the addresses pointed by some of
them.  This is incorrect, as the memory in question does not necessarily even
belong to the local SMPs.

Signed-off-by: Iliyan Malchev <malchev@google.com>