OMAP4: HDMI: Reprogram video wrapper after wrapper reset
For loading HDCP keys HDMI wrapper is reset before HDMI is powered
on either during an hpd 1 or set_mode when hdmi is being enabled.
HDCP authetication is started only when the
HDMI wrapper reset is completed and HDCP keys are loaded.

Also, when changing the HDMI modes, avoid HDCP recovery by
turning off HDCP in hdmi_power_off, instead of only for an hpd 0
interrupt.

Change-Id: I9a683f7a671c22969cbeed4d12d10bbb77d9d109
Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>