omap: rpmsg: unmap vring buffers after release mailbox
If the buffer is unmapped first there could be a message send from
the remoteproc and a kick (mailbox message) in that case the host will
access to the vring region in order to read the message but that region
is already unmaped. Now, the order is:

omap_mbox_put
rproc_put
iounmap

Which is the right order to do it.

Change-Id: Ifad91199fbdfdcd9228698ca2e32a005380e4802
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>