OMAP4+: EMIF: fix ddr self refresh cycles computation
ns_2_cycles uses the global variables T_num and T_den to convert
nanoseconds to cycles, but T_num and T_den are not kept up to date
on every memory bus frequency change.  Instead, they are updated only
when a memory frequency is used for the first time.

If ns_2_cycles is called in setup_registers, which runs on every
memory bus frequency change, it will use the memory bus frequency
that was most recently used for the first time.  Move the call to
ns_2_cycles into emif_calculate_regs, which is only called the
first time a memory bus frequency is used, and cache the result
for use in setup_registers.

Change-Id: I1db501acee73bcbd738d9fefc65563896ec76585
Signed-off-by: Colin Cross <ccross@android.com>